When user selects -d (disable encrypt) option, still encrypt, but use built-in and exposed special password
The security is the same as if the content were stored in
plaintext. But the 3rd party API seems to reject non-encrypted
payloads from file input as a bad request. This change makes the flam
compatible in both encrypt (default) mode or with crypto disabled.

This is a temporary fix -- until the 3rd party's url shortener API
changes again.